Added checks if constraints met for check_termination function and added nlopt tests to reference our implementation
Optimization previously terminated when constraints weren't met, now the first test case converges to the right solution when near the origin (hooray!).
Subsequent test cases are not so lucky. Moving the initial condition far away from the minimum also causes it to fail.

This could be due to several things, and the way nlopt handles these errors is illuminating.
Nlopt adds significant slack to the solution (e.g. 0.003 instead of 0), but meets both constraints. I'll seek to emulate this behavior in the next update.
This could be due to limiting the derivative to some finite large number or explicitly updating the tolrance (or more likely some more intelligent third solution).
Maybe a small constant is added to the offending variable, then the algorithm keeps running.
All solutions should meet constraints regardless, so if this isn't an issue dealing with operations on infinities it means there's something else wrong.